refactor: use mail instead email
