+++ Dockerfile
@@ -1,8 +1,8 @@
 FROM node:18
-WORKDIR /usr/src/email-service
-COPY email-service/package*.json ./
+WORKDIR /usr/src/mail-service
+COPY mail-service/package*.json ./
 RUN npm i
-COPY email-service/ .
+COPY mail-service/ .
 RUN npm run build
 RUN chmod +x ./start.sh
 CMD ["./start.sh"]
