+++ Dockerfile.prd
@@ -1,5 +1,5 @@
 FROM node:18
-WORKDIR /usr/src/auth-service
+WORKDIR /usr/src/mail-service
 COPY package*.json ./
 RUN npm i
 COPY . .
