+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mail-service",
-  "version": "1.1.5",
+  "version": "1.0.0",
   "description": "This project is a Medical Email Notification System designed to streamline communication between healthcare providers (doctors) and patients.",
   "main": "dist/main/server.js",
   "scripts": {
