+++ src/infrastructure/gateways/MailGateway.ts
@@ -1,10 +1,10 @@
 import { emailAdapter } from '@/main/adapters'
 
-export interface IEmailGateway {
+export interface IMailGateway {
   send: (to: string, subject: string, text: string, html: string) => Promise<void>
 }
 
-export class EmailGateway implements IEmailGateway {
+export class MailGateway implements IMailGateway {
   async send (to: string, subject: string, text: string, html: string): Promise<void> {
     await emailAdapter.send({ to, subject, text, html })
   }
