+++ src/infrastructure/handlers/AppointmentCreatedHandler.ts
@@ -1,8 +1,8 @@
 import { type Handler } from '@/infrastructure'
-import { type ISendEmail } from '@/usecases'
+import { type ISendMail } from '@/usecases'
 
 export class AppointmentCreatedHandler implements Handler {
-  constructor (private readonly _usecase: ISendEmail) { }
+  constructor (private readonly _usecase: ISendMail) { }
   async handle (event: any): Promise<void> {
     try {
       await this._usecase.execute(event)
