+++ src/main/factories/handlers/makeAppointmentCreatedHandler.ts
@@ -1,9 +1,9 @@
 import { type Handler, AppointmentCreatedHandler } from '@/infrastructure'
-import { EmailGateway } from '@/infrastructure/gateways/EmailGateway'
-import { SendEmail } from '@/usecases'
+import { MailGateway } from '@/infrastructure/gateways/MailGateway'
+import { SendMail } from '@/usecases'
 
 export const makeAppointmentCreatedHandler = (): Handler => {
-  const gateway = new EmailGateway()
-  const usecase = new SendEmail(gateway)
+  const gateway = new MailGateway()
+  const usecase = new SendMail(gateway)
   return new AppointmentCreatedHandler(usecase)
 }
