+++ tests/main/adapters/MailAdapter.spec.ts
